Clearly define the internal API for platform definitions
