+++ core/nix/src/internal/TzdbOnFilesystem.kt
@@ -5,12 +5,12 @@
 
 package kotlinx.datetime.internal
 
-internal class TzdbOnFilesystem(defaultTzdbPath: Path) {
+internal class TzdbOnFilesystem(defaultTzdbPath: Path): TimeZoneDatabase {
 
-    internal fun rulesForId(id: String): TimeZoneRules =
+    override fun rulesForId(id: String): TimeZoneRules =
         readTzFile(tzdbPath.resolve(Path.fromString(id)).readBytes()).toTimeZoneRules()
 
-    internal fun availableTimeZoneIds(): Set<String> = buildSet {
+    override fun availableTimeZoneIds(): Set<String> = buildSet {
         tzdbPath.traverseDirectory(exclude = tzdbUnneededFiles) { add(it.toString()) }
     }
